pyproject.toml
[project]
name = "xmr-cheque-bot"
version = "0.1.0"
description = "XMR Cheque Bot MVP — payment cheques with Monero"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "XMR Cheque Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "aiogram>=3.0.0",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "cryptography>=41.0.0",
    "structlog>=23.0.0",
    "qrcode>=7.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
xmr-bot = "xmr_cheque_bot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xmr_cheque_bot"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "W293", "B904", "F841", "SIM105", "SIM117"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E402"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_unused_configs = true
show_error_codes = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]